+++ include/ShaderState.hpp
@@ -1,12 +1,14 @@
 #pragma once
 
 #include <SDL3/SDL.h>
+
 #ifdef KRAKEN_ENABLE_PYTHON
 #include <nanobind/nanobind.h>
 #endif  // KRAKEN_ENABLE_PYTHON
 
 #include <filesystem>
 #include <string>
+#include <type_traits>
 
 #ifdef KRAKEN_ENABLE_PYTHON
 namespace nb = nanobind;
@@ -26,13 +28,20 @@ void _quit();
 class ShaderState
 {
   public:
-    ShaderState(const std::filesystem::path& fragmentFilePath, Uint32 uniformBufferCount = 0);
+    ShaderState(const std::filesystem::path& fragmentFilePath, uint32_t uniformBufferCount = 0);
     ~ShaderState();
 
     void bind() const;
     void unbind() const;
 
-    void setUniform(const Uint32 binding, const void* data, const size_t size) const;
+    template <typename UniformType>
+    void setUniform(const uint32_t binding, const UniformType& data) const
+    {
+        static_assert(
+            std::is_trivially_copyable_v<UniformType>, "Uniform data must be trivially copyable"
+        );
+        SDL_SetGPURenderStateFragmentUniforms(m_renderState, binding, &data, sizeof(UniformType));
+    }
 
   private:
     SDL_GPUShader* m_fragShader;
@@ -40,5 +49,8 @@ class ShaderState
 
     // Allow shader_state namespace to access private members for cleanup
     friend void shader_state::_quit();
+
+    // binding thing
+    friend void shader_state::_bind(nb::module_& module);
 };
 }  // namespace kn