+++ src/color.cpp
@@ -154,26 +154,6 @@ Color& Color::operator/=(const double scalar)
     return *this;
 }
 
-const Color Color::BLACK = {0, 0, 0};
-const Color Color::WHITE = {255, 255, 255};
-const Color Color::RED = {255, 0, 0};
-const Color Color::GREEN = {0, 255, 0};
-const Color Color::BLUE = {0, 0, 255};
-const Color Color::YELLOW = {255, 255, 0};
-const Color Color::MAGENTA = {255, 0, 255};
-const Color Color::CYAN = {0, 255, 255};
-const Color Color::GRAY = {128, 128, 128};
-const Color Color::DARK_GRAY = {64, 64, 64};
-const Color Color::LIGHT_GRAY = {192, 192, 192};
-const Color Color::ORANGE = {255, 165, 0};
-const Color Color::BROWN = {139, 69, 19};
-const Color Color::PINK = {255, 192, 203};
-const Color Color::PURPLE = {128, 0, 128};
-const Color Color::NAVY = {0, 0, 128};
-const Color Color::TEAL = {0, 128, 128};
-const Color(Color::OLIVE) = {128, 128, 0};
-const Color(Color::MAROON) = {128, 0, 0};
-
 namespace color
 {
 Color fromHex(std::string_view hex)